extracting PostsService interface to another file
