+++ main/mocks/posts.service.ts
@@ -1,7 +1,7 @@
 import { PostDto } from '../src/models/dto/Post.dto';
 import { PostModel } from '../src/models/response/Post.model';
 import { Observable, of } from 'rxjs';
-import { PostsServiceInterface } from '../src/posts/posts.service';
+import { PostsServiceInterface } from '../src/interfaces/PostsService.interface';
 
 export class PostsService implements PostsServiceInterface {
     private _posts: PostModel[] = [];
